# gateway-core/src/api/assets/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Label 309 Gateway API",
    "version": "1.0.0",
    "description": "The Label 309 Proof-of-Existence gateway HTTP API: quote, publish (single and batch), upload content for storage, list and retrieve on-chain records (everything a standalone verifier needs), read the account balance, and stream record/balance events.\n\nMost mutating endpoints require a Bearer API key carrying the relevant scope; the record collection and single-record reads are callable anonymously. The bearer key prefix, the error documentation base URL (the problem `type` member), and the server URLs are operator-configured per deployment.",
    "license": {
      "name": "Apache-2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "This deployment's data-plane base (the origin is operator-configured; the version segment is fixed)."
    }
  ],
  "paths": {
...
    "/poe/quote": {
      "post": {
        "operationId": "poe.quote",
        "tags": ["PoE"],
        "summary": "Lock the price for an upcoming /publish call. Returns a quote_id valid for 15 minutes.",
        "description": "Computes the cost of an upcoming publish from current FX (ADA/USD, AR/USD/byte, Cardano protocol params) plus the account margin, persists a publish_quote row, and returns the `quote_id` plus a breakdown of network + storage + service margin (USD-micros each, decimal strings to preserve bigint precision over JSON). The quote is valid for 15 minutes; POST /api/v1/poe/publish carries the `quote_id` and the server consumes it atomically with the balance debit. The response includes `fx_age_seconds`, the age of the underlying FX snapshot at quote time — the server does NOT gate on this; callers may implement their own age tolerance if needed. Quoting fails with 503 service-unavailable only when no FX rate snapshot exists at all (a cold-start pathology). Scope required: poe:create.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PoeQuoteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "200 — quote issued.",
            "headers": {
              "RateLimit-Limit": {
                "$ref": "#/components/headers/RateLimit-Limit"
              },
              "RateLimit-Remaining": {
                "$ref": "#/components/headers/RateLimit-Remaining"
              },
              "RateLimit-Reset": {
                "$ref": "#/components/headers/RateLimit-Reset"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PoeQuoteResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidBodyResponse"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedResponse"
          },
          "402": {
            "$ref": "#/components/responses/StorageFundingRefusedResponse"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenResponse"
          },
          "429": {
            "$ref": "#/components/responses/RateLimitedResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalErrorResponse"
          },
          "503": {
            "$ref": "#/components/responses/FxStaleResponse"
          }
        }
      }
    },
    "/poe/publish": {
      "post": {
        "operationId": "poe.publish",
        "tags": ["PoE"],
        "summary": "Publish a finalised Label 309 record to Cardano. JSON only; storage URIs must already be substituted.",
        "description": "Accepts a finalised canonical-CBOR record (hex) plus the `quote_id` from a prior POST /api/v1/poe/quote, and submits the record to Cardano under metadata label 309. The record MUST already carry its final `ar://` (or `ipfs://`, when supported) URIs: files are uploaded first via POST /api/v1/poe/uploads and the returned URIs belong in the record before it is published; a structurally invalid record is rejected with 422 `validation-failed`. Dedups by `sha256(record_bytes)` per account so a retry of the exact same record returns the existing resource (200) without spawning a duplicate Cardano submission and without consuming the quote. New submissions return 202 with `id`, `status: \"submitting\"`, and `balance_after_usd_micros` (decimal string). The Cardano `tx_hash` is populated asynchronously by the cardano-confirm worker; subscribe to GET /api/v1/poe/events/{poe_id} for live updates. The 402 problem document embeds `balance_usd_micros` and `required_usd_micros` (decimal strings, micro-USD) as extension members so the client can show the payer exactly how short the account is. Scope required: poe:create.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 255
            },
            "required": false,
            "name": "Idempotency-Key",
            "in": "header"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PoePublishRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "200 — dedup hit; existing resource returned. No balance change.",
            "headers": {
              "RateLimit-Limit": {
                "$ref": "#/components/headers/RateLimit-Limit"
              },
              "RateLimit-Remaining": {
                "$ref": "#/components/headers/RateLimit-Remaining"
              },
              "RateLimit-Reset": {
                "$ref": "#/components/headers/RateLimit-Reset"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PoePublishResponse"
                }
              }
            }
          },
          "202": {
            "description": "202 — new record accepted; resource `id` assigned, tx_hash pending. Returns `balance_after_usd_micros` as a decimal string.",
            "headers": {
              "RateLimit-Limit": {
                "$ref": "#/components/headers/RateLimit-Limit"
              },
              "RateLimit-Remaining": {
                "$ref": "#/components/headers/RateLimit-Remaining"
              },
              "RateLimit-Reset": {
                "$ref": "#/components/headers/RateLimit-Reset"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PoePublishResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidBodyResponse"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedResponse"
          },
          "402": {
            "$ref": "#/components/responses/InsufficientFundsResponse"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenResponse"
          },
          "404": {
            "$ref": "#/components/responses/QuoteNotFoundResponse"
          },
          "409": {
            "$ref": "#/components/responses/QuoteAlreadyConsumedResponse"
          },
          "410": {
            "$ref": "#/components/responses/QuoteExpiredResponse"
          },
          "422": {
            "$ref": "#/components/responses/ValidationFailedResponse"
          },
          "429": {
            "$ref": "#/components/responses/RateLimitedResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalErrorResponse"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque",
        "description": "API key authentication. Pass the operator-issued API key in the Authorization: Bearer header. The key's human-readable prefix and which deployment (network) it is scoped to are operator-configured."
      }
    },
    "schemas": {
      "PoePublishResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^poe_[0-9a-hjkmnp-tv-z]{26}$"
          },
          "tx_hash": {
            "type": ["string", "null"],
            "pattern": "^[0-9a-f]{64}$"
          },
          "status": {
            "type": "string",
            "enum": ["submitting", "submitted", "confirmed", "permanent_failure"]
          },
          "items_count": {
            "type": "integer",
            "minimum": 0
          },
          "signed": {
            "type": "boolean"
          },
          "sealed": {
            "type": "boolean"
          },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "item_idx": {
                  "type": "integer",
                  "minimum": 0
                },
                "hashes": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string",
                    "pattern": "^[0-9a-f]+$"
                  }
                },
                "uris": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "enc": {
                  "type": "object",
                  "properties": {
                    "scheme": {
                      "type": "integer"
                    },
                    "aead": {
                      "type": "string"
                    },
                    "kem": {
                      "type": "string"
                    },
                    "nonce": {
                      "type": "string",
                      "pattern": "^[0-9a-f]+$"
                    },
                    "slots_count": {
                      "type": "integer",
                      "minimum": 0
                    },
                    "slots_mac": {
                      "type": "string",
                      "pattern": "^[0-9a-f]+$"
                    },
                    "passphrase": {
                      "type": "object",
                      "properties": {
                        "alg": {
                          "type": "string"
                        },
                        "salt": {
                          "type": "string",
                          "pattern": "^[0-9a-f]+$"
                        },
                        "params": {
                          "type": "object",
                          "additionalProperties": {}
                        }
                      },
                      "required": ["alg", "salt", "params"]
                    }
                  },
                  "additionalProperties": false
                }
              },
              "required": ["item_idx", "hashes"],
              "additionalProperties": false
            }
          },
          "conformance_profile": {
            "type": "string",
            "enum": ["core", "signed", "sealed", "recipient-sealed"]
          },
          "balance_after_usd_micros": {
            "type": "string",
            "pattern": "^[0-9]+$"
          }
        },
        "required": [
          "id",
          "tx_hash",
          "status",
          "items_count",
          "signed",
          "sealed",
          "items",
          "conformance_profile",
          "balance_after_usd_micros"
        ],
        "additionalProperties": false
      },
...
      "PoeQuoteRequest": {
        "type": "object",
        "properties": {
          "record_bytes": {
            "type": "integer",
            "exclusiveMinimum": 0
          },
          "recipient_count": {
            "type": "integer",
            "minimum": 0
          },
          "file_bytes_total": {
            "type": "integer",
            "minimum": 0
          }
        },
        "required": ["record_bytes", "recipient_count", "file_bytes_total"],
        "additionalProperties": false
      },
...
      "PoePublishRequest": {
        "type": "object",
        "properties": {
          "record": {
            "type": "string",
            "pattern": "^[0-9a-f]+$"
          },
          "quote_id": {
            "type": "string",
            "format": "uuid"
          },
          "signatures": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "cose_sign1": {
                  "type": "string",
                  "pattern": "^[0-9a-f]+$"
                },
                "cose_key": {
                  "type": "string",
                  "pattern": "^[0-9a-f]+$"
                }
              },
              "required": ["cose_sign1"],
              "additionalProperties": false
            }
          }
        },
        "required": ["record", "quote_id"],
        "additionalProperties": false
      },
      "PoeQuoteResponse": {
        "type": "object",
        "properties": {
          "quote_id": {
            "type": "string",
            "format": "uuid"
          },
          "usd_micros": {
            "type": "string",
            "pattern": "^[0-9]+$"
          },
          "breakdown": {
            "type": "object",
            "properties": {
              "network_usd_micros": {
                "type": "string",
                "pattern": "^[0-9]+$"
              },
              "storage_usd_micros": {
                "type": "string",
                "pattern": "^[0-9]+$"
              },
              "service_usd_micros": {
                "type": "string",
                "pattern": "^[0-9]+$"
              }
            },
            "required": ["network_usd_micros", "storage_usd_micros", "service_usd_micros"],
            "additionalProperties": false
          },
          "margin_pct": {
            "type": "number"
          },
          "margin_source": {
            "type": "string",
            "description": "The attribution of the applied margin: 'account-override' when a per-account override priced the quote, else 'operator-default'."
          },
          "expires_at": {
            "type": "string",
            "format": "date-time"
          },
          "fx_age_seconds": {
            "type": "integer",
            "minimum": 0
          },
          "amount": {
            "type": "string",
            "pattern": "^[0-9]+$",
            "description": "The locked total as a decimal string of USD micro-cents (the SDK promotes it to an arbitrary-precision integer at the application boundary)."
          },
          "currency": {
            "type": "string",
            "description": "ISO 4217 currency the amount is denominated in.",
            "example": "USD"
          }
        },
        "required": [
          "quote_id",
          "amount",
          "currency",
          "usd_micros",
          "breakdown",
          "margin_pct",
          "expires_at",
          "fx_age_seconds"
        ],
        "additionalProperties": false
      },
...
    },
    "responses": {
      "InvalidBodyResponse": {
        "description": "400 — invalid-body",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            },
            "example": {
              "type": "https://gateway.example/api/v1/errors#invalid-body",
              "title": "Invalid Body",
              "status": 400,
              "detail": "Request body is not valid JSON.",
              "code": "invalid-body",
              "trace_id": "01977c8a-7c5e-7000-8000-000000000000"
            }
          }
        }
      },
...
      "QuoteNotFoundResponse": {
        "description": "404 — quote-not-found",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            },
            "example": {
              "type": "https://gateway.example/api/v1/errors#quote-not-found",
              "title": "Quote Not Found",
              "status": 404,
              "detail": "No publish quote with that id was found for this account.",
              "code": "quote-not-found",
              "trace_id": "01977c8a-7c5e-7000-8000-000000000000"
            }
          }
        }
      },
...
      "QuoteAlreadyConsumedResponse": {
        "description": "409 — quote-already-consumed",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            },
            "example": {
              "type": "https://gateway.example/api/v1/errors#quote-already-consumed",
              "title": "Quote Already Consumed",
              "status": 409,
              "detail": "This publish quote has already been consumed; request a new quote and retry.",
              "code": "quote-already-consumed",
              "trace_id": "01977c8a-7c5e-7000-8000-000000000000"
            }
          }
        }
      },
      "InsufficientFundsResponse": {
        "description": "402 — insufficient-funds. The problem body carries the RFC 7807 §3.2 extension members `balance_usd_micros` (the balance at rejection time) and `required_usd_micros` (the charge it could not cover), both decimal strings in micro-USD.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            },
            "example": {
              "type": "https://gateway.example/api/v1/errors#insufficient-funds",
              "title": "Insufficient Funds",
              "status": 402,
              "detail": "The account does not have enough USD balance to publish the record(s).",
              "code": "insufficient-funds",
              "balance_usd_micros": "150000",
              "required_usd_micros": "410000",
              "trace_id": "01977c8a-7c5e-7000-8000-000000000000"
            }
          }
        }
      },
...
      "ForbiddenResponse": {
        "description": "403 — insufficient-scope. The problem body carries the RFC 7807 §3.2 extension members `required` (string[]: the scope the endpoint requires) and `granted` (string[]: the scopes the credential actually carries — empty for an operator credential presented to the data plane, which grants nothing there).",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            },
            "example": {
              "type": "https://gateway.example/api/v1/errors#insufficient-scope",
              "title": "Insufficient Scope",
              "status": 403,
              "detail": "this endpoint requires the poe:create scope",
              "code": "insufficient-scope",
              "required": ["poe:create"],
              "granted": ["poe:read"],
              "trace_id": "01977c8a-7c5e-7000-8000-000000000000"
            }
          }
        }
      },
      "UnauthorizedResponse": {
        "description": "401 — unauthorized",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            },
            "example": {
              "type": "https://gateway.example/api/v1/errors#unauthorized",
              "title": "Unauthorized",
              "status": 401,
              "detail": "Authentication is required to access this resource. The `code` field is always `unauthorized`: a missing Authorization header, a malformed Bearer credential, and an unknown or revoked one all collapse to the same response, so a scanner cannot distinguish credential formats from live secrets.",
              "code": "unauthorized",
              "trace_id": "01977c8a-7c5e-7000-8000-000000000000"
            }
          }
        }
      },
...
      "InternalErrorResponse": {
        "description": "500 — internal-error",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            },
            "example": {
              "type": "https://gateway.example/api/v1/errors#internal-error",
              "title": "Internal Error",
              "status": 500,
              "detail": "An unexpected server error occurred.",
              "code": "internal-error",
              "trace_id": "01977c8a-7c5e-7000-8000-000000000000"
            }
          }
        }
      },
      "QuoteExpiredResponse": {
        "description": "410 — quote-expired",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            },
            "example": {
              "type": "https://gateway.example/api/v1/errors#quote-expired",
              "title": "Quote Expired",
              "status": 410,
              "detail": "The publish quote has expired; request a new quote and retry.",
              "code": "quote-expired",
              "trace_id": "01977c8a-7c5e-7000-8000-000000000000"
            }
          }
        }
      },
...
      "RateLimitedResponse": {
        "description": "429 — rate-limited",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            },
            "example": {
              "type": "https://gateway.example/api/v1/errors#rate-limited",
              "title": "Rate Limited",
              "status": 429,
              "detail": "API-key rate limit exceeded.",
              "code": "rate-limited",
              "trace_id": "01977c8a-7c5e-7000-8000-000000000000"
            }
          }
        }
      },
      "FxStaleResponse": {
        "description": "503 — service-unavailable",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            },
            "example": {
              "type": "https://gateway.example/api/v1/errors#service-unavailable",
              "title": "Service Unavailable",
              "status": 503,
              "detail": "No FX rate snapshot exists; pricing is unavailable. The worker may still be booting; please retry shortly.",
              "code": "service-unavailable",
              "trace_id": "01977c8a-7c5e-7000-8000-000000000000"
            }
          }
        }
      },
      "StorageFundingRefusedResponse": {
        "description": "402 — the storage funding the publish would draw cannot be honoured. The `code` is `no-funding-grant` when no live grant entitles the caller to draw a funding source for the configured backend, or `insufficient-storage-credit` when the drawable source's cached credit is below the safety floor or its provider-reported capacity cannot cover the chargeable bytes.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            },
            "example": {
              "type": "https://gateway.example/api/v1/errors#no-funding-grant",
              "title": "No storage funding grant",
              "status": 402,
              "detail": "No storage funding source entitles this account to store content beyond the free window.",
              "code": "no-funding-grant",
              "trace_id": "01977c8a-7c5e-7000-8000-000000000000"
            }
          }
        }
      },
...
    },
    "headers": {
      "RateLimit-Reset": {
        "description": "IETF draft-ietf-httpapi-ratelimit-headers-09. Seconds until the current sliding window rolls over — the earliest moment `RateLimit-Remaining` could fully reset.",
        "schema": {
          "type": "integer",
          "minimum": 1
        }
      },
      "RateLimit-Remaining": {
        "description": "IETF draft-ietf-httpapi-ratelimit-headers-09. Tokens left in the current sliding window after this request was counted. Zero on a 429 response.",
        "schema": {
          "type": "integer",
          "minimum": 0
        }
      },
      "RateLimit-Limit": {
        "description": "IETF draft-ietf-httpapi-ratelimit-headers-09. The per-window request ceiling enforced for the API key on this endpoint, in requests per 60 seconds.",
        "schema": {
          "type": "integer",
          "minimum": 1
        }
      }
    }
  }
}